% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backstep.R
\name{backstep}
\alias{backstep}
\title{Backstep procedure}
\usage{
backstep(h_fim, d, s, n, dec, coef_em, exp_em, rc)
}
\arguments{
\item{h_fim}{Pressure at the end of the line in mca}

\item{d}{Pipe diameter in meters}

\item{s}{Emitter spacing in meters}

\item{n}{Number of emitters}

\item{dec}{Slope on the lateral line}

\item{coef_em}{Emission coefficient}

\item{exp_em}{Emission exponent}

\item{rc}{Roughness coefficient. Absolute roughness (in meters) for
Universal equation.}
}
\description{
\code{backstep} Calculate the pressure head and flow rate profile along
a lateral line
}
\examples{
backstep(h_fim = 10, d = 0.012, s = 1, n = 5, dec = 0, coef_em = 8.84e-7,
exp_em = 0.50, rc = 2e-6)
backstep(h_fim = 10, d = 0.012, s = c(0.3, 0.3, 1), n = 12, dec = 0,
coef_em = 8.84e-7, exp_em = 0.50, rc = 2e-6)
}
