% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection}
\alias{bisection}
\title{Bisection}
\usage{
bisection(f, a = 100, b = 1e-10, i_max = 100, tol = 1e-10)
}
\arguments{
\item{f}{function}

\item{a}{max}

\item{b}{min}

\item{i_max}{max iterations}

\item{tol}{Convergence tolerance}
}
\description{
Bisection
}
