% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_emitter.R
\name{coef_emitter}
\alias{coef_emitter}
\title{Emmiter equation}
\usage{
coef_emitter(h, q)
}
\arguments{
\item{h}{Pressure}

\item{q}{Flow rate}
}
\description{
\code{coef_emitter} Calculates the coefficients of the emitter pressure-flow power equation
}
\examples{
h <- c(10, 20, 30)
q <- c(7.9, 11.4, 14.1)
coef_emitter(h, q)
}
