% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_lateral.R
\name{coef_lateral}
\alias{coef_lateral}
\title{Lateral coefficients}
\usage{
coef_lateral(
  d_lateral,
  s_lateral,
  s_ini_lateral = s_lateral,
  n_lateral,
  dec_lateral,
  coef_em,
  exp_em,
  rc,
  sim = list(1, 100, 1)
)
}
\arguments{
\item{d_lateral}{Diameter of the lateral line, meter}

\item{s_lateral}{Spacing between emitters on the lateral line, meter}

\item{s_ini_lateral}{Spacing of the initial stretch of the lateral line, meter}

\item{n_lateral}{Number of issuers on the lateral line}

\item{dec_lateral}{Lateral line slope, decimal}

\item{coef_em}{Coefficient of the flow-pressure equation of the emitter}

\item{exp_em}{Exponent of the flow-pressure equation of the emitter}

\item{rc}{Absolute roughness relative to pipe material, meter}

\item{sim}{Limites da simulação}
}
\description{
\code{coef_lateral} calculates lateral line coefficients
}
\examples{
args_coef_lateral <- list(
  d_lateral = 0.012, s_lateral = 0.20, s_ini_lateral = 1,
  n_lateral = 200, dec_lateral = -0.01, coef_em = 1.67e-7, exp_em = 0.52, rc = 1e-4
)
do.call(coef_lateral, args_coef_lateral)
}
