% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefic.R
\name{coefic}
\alias{coefic}
\title{Coefficient}
\usage{
coefic(K, x, D, Se, ne, dec, HL = c(1, 100, 0.5))
}
\arguments{
\item{K}{Emission coefficient}

\item{x}{Emission exponent}

\item{D}{Diameter in meters}

\item{Se}{Emitter spacing in meters}

\item{ne}{Number of emitters}

\item{dec}{Slope on the lateral line}
}
\value{
coefficient a, b e c
}
\description{
Calculate the coefficients of lateral line
}
\examples{
coefic( K = 1.053e-6, x = 0.5, D = 0.015, Se = 1, ne = 100, dec = 0.00, HL = c(1, 100, 0.5))
coefic( K = 1.053e-6, x = 0.5, D = 0.015, Se = 1, ne = 100, dec = -0.01, HL = c(1, 100, 0.5))
}
