% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0.R
\name{et0_calc}
\alias{et0_calc}
\title{Reference Evapotranspiration - The FAO Penman-Monteith method}
\usage{
et0_calc(
  tem_max,
  tem_min,
  umd_max,
  umd_min,
  uv,
  uz = 2,
  rs,
  g = 0,
  lat,
  alt,
  date
)
}
\arguments{
\item{tem_max}{Maximum air temperature, C}

\item{tem_min}{Minimum air temperature, C}

\item{umd_max}{Maximum relative humidity, \%}

\item{umd_min}{Minimum relative humidity, \%}

\item{uv}{Wind speed, m/s}

\item{uz}{Height of measurement of wind speed above ground surface. Default is 2 m.}

\item{rs}{Solar radiation, MJ m^-2^^day^-1}

\item{g}{Soil heat flux. Default is 0}

\item{lat}{Latitude,}

\item{alt}{Altitude abose sea level}

\item{date}{}
}
\value{

}
\description{
Reference Evapotranspiration - The FAO Penman-Monteith method
}
\examples{
et0_calc(
  tem_max = 21.5, tem_min = 12.3, umd_max = 84, umd_min = 63, uv = 2.78, uz = 10,
  rs = 22.07, g = 0, lat = 50.8, alt = 100, date = as.Date("2009-7-6")
)

}
