% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateral_diameter.R
\name{lateral_diameter}
\alias{lateral_diameter}
\title{Diameter of the lateral line}
\usage{
lateral_diameter(
  q_var_lateral = 10,
  h_fim_lateral,
  s_lateral,
  s_ini_lateral = s_lateral,
  n_lateral,
  dec_lateral,
  coef_em,
  exp_em,
  q_unit = q_unit
)
}
\arguments{
\item{q_var_lateral}{Lateral maximum variation in flow rate, percentage}

\item{h_fim_lateral}{Pressure at the end of the lateral line, mca}

\item{s_lateral}{Spacing between emitters on the lateral line, meter}

\item{s_ini_lateral}{Spacing of the initial stretch of the lateral line, meter}

\item{n_lateral}{Number of issuers on the lateral line}

\item{dec_lateral}{Lateral line slope, decimal}

\item{coef_em}{Coefficient of the flow-pressure equation of the emitter}

\item{exp_em}{Exponent of the flow-pressure equation of the emitter}

\item{q_unit=q_unit}{aditional parameters}
}
\description{
\code{lateral_q_req} calculates the diameter of the lateral line to achieve
a given variation in flow rate
}
\examples{
args_lateral_diameter <- list(
  q_var_lateral = 10, h_fim_lateral = 10, s_lateral = 0.20, s_ini_lateral = 1,
  n_lateral = 200, dec_lateral = -0.01, coef_em = 1.67e-7*360000, exp_em = 0.52, q_unit="l/h"
)
do.call(lateral_diameter, args_lateral_diameter)
}
