% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateral_head.R
\name{lateral_head}
\alias{lateral_head}
\title{Pressure required on the lateral line}
\usage{
lateral_head(
  q_req_lateral,
  d_lateral,
  s_lateral,
  s_ini_lateral = s_lateral,
  n_lateral,
  dec_lateral,
  coef_em,
  exp_em,
  q_unit = q_unit
)
}
\arguments{
\item{q_req_lateral}{Average flow required on the lateral line, m3/s}

\item{d_lateral}{Diameter of the lateral line, milimeter}

\item{s_lateral}{Spacing between emitters on the lateral line, meter}

\item{s_ini_lateral}{Spacing of the initial stretch of the lateral line, meter}

\item{n_lateral}{Number of issuers on the lateral line}

\item{dec_lateral}{Lateral line slope, decimal}

\item{coef_em}{Coefficient of the flow-pressure equation of the emitter}

\item{exp_em}{Exponent of the flow-pressure equation of the emitter}

\item{q_unit=q_unit}{additional parameters}
}
\description{
\code{lateral_head} Calculates the pressure at the beginning of the lateral
line for a given average flow rate
}
\examples{
args_lateral_head <- list(
  q_req_lateral = 5.5555e-7*3600000
  ,
   d_lateral = 12
   ,
    s_lateral = 0.20
    ,
     s_ini_lateral = 1
     ,

  n_lateral = 200
  ,
   dec_lateral = -0.00
   ,
    coef_em = 1.67e-7*3600000
    ,
     exp_em = 0.52
     ,
      q_unit="l/h"
)
do.call(lateral_head, args_lateral_head)
}
