% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateral_head_diameter.R
\name{lateral_head_diameter}
\alias{lateral_head_diameter}
\title{Pressure and diameter on the lateral line}
\usage{
lateral_head_diameter(
  q_req_lateral,
  q_var_lateral = 10,
  s_lateral,
  s_ini_lateral = s_lateral,
  n_lateral,
  dec_lateral,
  coef_em,
  exp_em,
  rc,
  ...
)
}
\arguments{
\item{q_req_lateral}{Average flow required on the lateral line, m3/s}

\item{q_var_lateral}{Lateral maximum variation in flow rate, percentage}

\item{s_lateral}{Spacing between emitters on the lateral line, meter}

\item{s_ini_lateral}{Spacing of the initial stretch of the lateral line, meter}

\item{n_lateral}{Number of issuers on the lateral line}

\item{dec_lateral}{Lateral line slope, decimal}

\item{coef_em}{Coefficient of the flow-pressure equation of the emitter}

\item{exp_em}{Exponent of the flow-pressure equation of the emitter}

\item{rc}{Absolute roughness relative to pipe material, meter}

\item{...}{additional parameters}
}
\description{
\code{lateral_head_diameter} calculates lateral line diameter and
pressure for a given average flow rate and a given flow rate variation
}
\examples{
args_lateral_head_diameter <- list(
  q_req_lateral = 5.5555e-7, q_var_lateral = 10, s_lateral = 0.20, s_ini_lateral = 1,
  n_lateral = 200, dec_lateral = -0.01, coef_em = 1.67e-7, exp_em = 0.52, rc = 1e-4
)
do.call(lateral_head_diameter, args_lateral_head_diameter)
}
