% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateral_profile.R
\name{lateral_profile}
\alias{lateral_profile}
\title{Pressure and flow profile in a lateral line}
\usage{
lateral_profile(
  h_fim_lateral,
  d_lateral,
  s_lateral,
  s_ini_lateral = s_lateral,
  n_lateral,
  dec_lateral,
  coef_em,
  exp_em,
  rc
)
}
\arguments{
\item{h_fim_lateral}{Pressure at the end of the lateral line, mca}

\item{d_lateral}{Diameter of the lateral line, meter}

\item{s_lateral}{Spacing between emitters on the lateral line, meter}

\item{s_ini_lateral}{Spacing of the initial stretch of the lateral line, meter}

\item{n_lateral}{Number of issuers on the lateral line}

\item{dec_lateral}{Lateral line slope, decimal}

\item{coef_em}{Coefficient of the flow-pressure equation of the emitter}

\item{exp_em}{Exponent of the flow-pressure equation of the emitter}

\item{rc}{Absolute roughness relative to pipe material, meter}
}
\description{
\code{lateral_profile} Calculates the pressure and flow profile on the lateral line
}
\examples{
args_lateral_profile <- list(
  h_fim_lateral = 10, d_lateral = 0.012, s_lateral = 0.20, s_ini_lateral = 1,
  n_lateral = 200, dec_lateral = -0.01, coef_em = 1.67e-7, exp_em = 0.52, rc = 1e-4)
do.call(lateral_profile, args_lateral_profile)
}
