% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backstep.R
\name{ll}
\alias{ll}
\title{Design a lateral line}
\usage{
ll(Hfim, K, x, D, Se, ne, dec)
}
\arguments{
\item{Hfim}{Pressure at the end of the line in mca}

\item{K}{Emission coefficient}

\item{x}{Emission exponent}

\item{D}{Diameter in meters}

\item{Se}{Emitter spacing in meters}

\item{ne}{Number of emitters}

\item{dec}{Slope on the lateral line}
}
\value{
Hini pressure head at inlet of lateral, Qini flow rate at inlet of the lateral, H pressure head in the emitter,  q flow rate in the emitter,
Q flow rate in lateral line, hf head loss
}
\description{
Calculate the pressure head and flow rate profile along a lateral line
}
\examples{
ll(Hfim = 10, K = 1.053e-6, x = 0.5, D = 0.025, Se = 1, ne = 10, dec = 0.00)
ll(Hfim = 10, K = 1.053e-6, x = 0.5, D = 0.025, Se = 1, ne = 10, dec = 0.01)
ll(Hfim = 10, K = 1.053e-6, x = 0.5, D = 0.025, Se = 1, ne = 10, dec = -0.01)
ll(Hfim = 10, K = 1.053e-6, x = 0.5, D = 0.025, Se = 1, ne = 100, dec = 0.00)
ll(Hfim = 10, K = 1.053e-6, x = 0.5, D = 0.025, Se = 1, ne = 100, dec = 0.01)
ll(Hfim = 10, K = 1.053e-6, x = 0.5, D = 0.025, Se = 1, ne = 100, dec = -0.01)
}
