% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lld.R
\name{lld}
\alias{lld}
\title{Adjust the emitter flow variation to change the diameter}
\usage{
lld(Hfim, K, x, Se, ne, dec, qvr, imax = 100, toler = 1e-06, DLLa = 1,
  DLLb = 0)
}
\arguments{
\item{Hfim}{Pressure at the end of the line in mca}

\item{K}{Emission coefficient}

\item{x}{Emission exponent}

\item{Se}{Emitter spacing in meters}

\item{ne}{Number of emitters}

\item{dec}{Slope on the lateral line}

\item{qvr}{Emitter flow variation}

\item{imax}{Maximum number of iteration}

\item{DLLa}{Maximum diameter in meters}

\item{DLLb}{Minimum diameter in meters}

\item{D}{Diameter in meters}
}
\value{
D diameter, qv emitter flow variation, iter number iterations, tempo time
}
\description{
Set the diameter of a lateral line
}
\examples{
lld(Hfim = 10, K = 1.053e-6, x = 0.5, Se = 1, ne = 100, dec =  0.00, qvr = 10, imax = 100, toler = 1e-6, DLLa = 1, DLLb = 0)
lld(Hfim = 10, K = 1.053e-6, x = 0.5, Se = 1, ne = 100, dec =  0.01, qvr = 10, imax = 100, toler = 1e-6, DLLa = 1, DLLb = 0)
lld(Hfim = 10, K = 1.053e-6, x = 0.5, Se = 1, ne = 100, dec = -0.01, qvr = 10, imax = 100, toler = 1e-6, DLLa = 1, DLLb = 0)
}
