% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llh.R
\name{llh}
\alias{llh}
\title{Establish the pressure at the start of a lateral line}
\usage{
llh(D, K, x, Se, ne, dec, qreq, imax = 100, toler = 1e-10, Hfima = 100,
  Hfimb = 0)
}
\arguments{
\item{D}{Diameter in meters}

\item{K}{Emission coefficient}

\item{x}{Emission exponent}

\item{Se}{Emitter spacing in meters}

\item{ne}{Number of emitters}

\item{dec}{Slope on the lateral line}

\item{qreq}{Flow rate required at the emitters in cubic meters per second}

\item{imax}{Maximum number of iteration}

\item{Hfima}{Maximum pressure head at the end of the lateral line in meters}

\item{Hfimb}{Minimum pressure head at the end of the lateral line in meters}
}
\value{
Hfim pressure head at the end of the lateral line, qsistema system flow rate, iter number iterations,tempo time
}
\description{
Determine the pressure at the start of the lateral line
}
\examples{
llh( D = 0.025, K = 1.053e-6, x = 0.5, Se = 1, ne = 10, dec = 0.0, qreq = 3.333333e-6, imax = 100, toler = 1e-10, Hfima = 30, Hfimb = 0)
llh( D = 0.025, K = 1.053e-6, x = 0.5, Se = 1, ne = 10, dec = 0.04, qreq = 3.333333e-6, imax = 100, toler = 1e-10,Hfima = 30, Hfimb = 0)
llh( D = 0.025, K = 1.053e-6, x = 0.5, Se = 1, ne = 10, dec = -0.04, qreq = 3.333333e-6, imax = 100, toler = 1e-10,Hfima = 30, Hfimb = 0)
}
