% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llhd.R
\name{llhd}
\alias{llhd}
\title{Established to set the diameter and pressure required at the start of a lateral line}
\usage{
llhd(D, K, x, Se, ne, dec, qreq, imax = 100, toler = 1e-06, qvr)
}
\arguments{
\item{D}{Diameter in meters}

\item{K}{Emission coefficient}

\item{x}{Emission exponent}

\item{Se}{Emitter spacing in meters}

\item{ne}{Number of emitters}

\item{dec}{Slope on the lateral line}

\item{qreq}{Flow rate required at the emitters in cubic meters per second}

\item{imax}{Maximum number of iteration}

\item{toler}{Maximum tolerance}

\item{qvr}{Emitter flow variation}
}
\value{
D diameter, Hfim pressure head at the end of the lateral line, iter number of iterations, tempo time
}
\description{
Determine a linha lateral e o diâmetro de pressão necessários no início de uma linha lateral
}
\examples{
llhd( D = 0.25, K = 1.053e-6, x = 0.5, Se = 1, ne = 10, dec = 0.00, qreq = 3.33333e-6, imax = 100, toler = 1e-10, qvr = 10)
llhd( D = 0.25, K = 1.053e-6, x = 0.5, Se = 1, ne = 10, dec = 0.04, qreq = 3.33333e-6, imax = 100, toler = 1e-10, qvr = 10)
llhd( D = 0.25, K = 1.053e-6, x = 0.5, Se = 1, ne = 10, dec = 0.04, qreq = 3.33333e-6, imax = 100, toler = 1e-10, qvr = 10)
}
