% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subuni.R
\name{subuni}
\alias{subuni}
\title{Define an irrigation subunit with lateral line and the manifold}
\usage{
subuni(HfimLD, K, x, DLL, DLD, SeLD, SeLL, nLL, ne, dec)
}
\arguments{
\item{HfimLD}{Pressure at the end of the manifold in mca}

\item{K}{Emission coefficient}

\item{x}{Emission exponent}

\item{DLL}{Diameter lateral line in meters}

\item{DLD}{Diameter manifold in meters}

\item{SeLD}{Lateral line spacing in meters}

\item{SeLL}{Emitter spacing in meters}

\item{ne}{Number of emitters in meters}

\item{dec}{Slope on the lateral line}

\item{neLL}{Number of lateral lines in meters}
}
\value{
h Pressure profile and q Flow rate Profile
}
\description{
calculate of the pressure head and flow rate profile a subunit
}
\examples{
subuni( HfimLD = 15, K = 1.053e-6, x = 0.5, DLL = 0.025, DLD = 0.040, SeLD = 4, SeLL = 1, nLL = 10, ne = 100, dec = 0.0)
}
