% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunihd.R
\name{subunihd}
\alias{subunihd}
\title{Established to set the diameter lateral line and pressure required at the start a subunit}
\usage{
subunihd(DLL, K, x, DLD, SeLL, SeLD, nLL, ne, dec, qreq, imax = 100,
  toler = 1e-06, qvr)
}
\arguments{
\item{DLL}{Diameter lateral line in meters}

\item{K}{Emission coefficient}

\item{x}{Emission exponent}

\item{DLD}{Diameter manifold in meters}

\item{SeLL}{Emitter spacing in meters}

\item{SeLD}{Lateral line spacing in meters}

\item{nLL}{Number of lateral lines in meters}

\item{ne}{Number of emitters in meters}

\item{dec}{Slope on the lateral line}

\item{qreq}{Flow rate required at the emitters in cubic meters per second}

\item{imax}{Maximum number of iteration}

\item{toler}{Maximum tolerance}

\item{qvr}{Emitter flow variation}
}
\value{
D diameter lateral line, pressure head at the end, iter number of iterations, tempo time
}
\description{
Set the diameter lateral line and pressure required at the start  a subunit
}
\examples{
subunihd(DLL = 0.025,K = 1.053e-6,x = 0.5,DLD = 0.04,
SeLL = 4, SeLD = 1, nLL = 100, ne = 10, dec = 0.00,
qreq = 3.33333e-6, imax = 100,toler = 1e-6, qvr = 10)
}
