% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunit_head_diameter.R
\name{subunit_head_diameter}
\alias{subunit_head_diameter}
\title{Pressure and diameter on the manifold line}
\usage{
subunit_head_diameter(
  q_req_subunit,
  q_var_subunit = 10,
  s_manifold,
  s_ini_manifold = s_manifold,
  n_manifold,
  dec_manifold,
  d_lateral,
  s_lateral,
  s_ini_lateral = s_lateral,
  n_lateral,
  dec_lateral,
  coef_em,
  exp_em,
  q_unit = q_unit
)
}
\arguments{
\item{q_req_subunit}{Average flow required on the subunit, m3/s}

\item{q_var_subunit}{Subunit maximum variation in flow rate, peq_unit=q_unitentage}

\item{s_manifold}{Spacing between lateral lines in the manifold line, meter}

\item{s_ini_manifold}{Spacing of the initial section of the manifold line, meter}

\item{n_manifold}{Number of lateral lines in the manifold line}

\item{dec_manifold}{Slope of the manifold line, decimal}

\item{d_lateral}{Diameter of the lateral line, milimeter}

\item{s_lateral}{Spacing between emitters on the lateral line, meter}

\item{s_ini_lateral}{Spacing of the initial stretch of the lateral line, meter}

\item{n_lateral}{Number of issuers on the lateral line}

\item{dec_lateral}{Lateral line slope, decimal}

\item{coef_em}{Coefficient of the flow-pressure equation of the emitter}

\item{exp_em}{Exponent of the flow-pressure equation of the emitter}
}
\description{
\code{subunit_head_diameter} calculates manifold line diameter and
pressure for a given average flow rate and a given flow rate variation
}
\examples{
args_subunit_head_diameter <- list(
  q_req_subunit = 2.2222e-6*3600000, q_var_subunit = 10,
  s_manifold = c(0.3, 0.3, 2), s_ini_manifold = 10, n_manifold = 12, dec_manifold = 0.02,
  d_lateral = 16, s_lateral = 1, s_ini_lateral = 4, n_lateral = 40, dec_lateral = -0.01,
  coef_em = 6.41e-7*3600000, exp_em = 0.54, q_unit="l/h"
)
do.call(subunit_head_diameter, args_subunit_head_diameter)
}
