% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunit_diameter.R
\name{subunit_diameter}
\alias{subunit_diameter}
\title{Diameter of the manifold line}
\usage{
subunit_diameter(
  q_var_subunit = 10,
  h_fim_manifold,
  s_manifold,
  s_ini_manifold = s_manifold,
  n_manifold,
  dec_manifold,
  d_lateral,
  s_lateral,
  s_ini_lateral = s_lateral,
  n_lateral,
  dec_lateral,
  coef_em,
  exp_em,
  rc,
  ...
)
}
\arguments{
\item{q_var_subunit}{Subunit maximum variation in flow rate, percentage}

\item{h_fim_manifold}{Pressure at the end of the manifold line, mca}

\item{s_manifold}{Spacing between lateral lines in the manifold line, meter}

\item{s_ini_manifold}{Spacing of the initial section of the manifold line, meter}

\item{n_manifold}{Number of lateral lines in the manifold line}

\item{dec_manifold}{Slope of the manifold line, decimal}

\item{d_lateral}{Diameter of the lateral line, meter}

\item{s_lateral}{Spacing between emitters on the lateral line, meter}

\item{s_ini_lateral}{Spacing of the initial stretch of the lateral line, meter}

\item{n_lateral}{Number of issuers on the lateral line}

\item{dec_lateral}{Lateral line slope, decimal}

\item{coef_em}{Coefficient of the flow-pressure equation of the emitter}

\item{exp_em}{Exponent of the flow-pressure equation of the emitter}

\item{rc}{Absolute roughness relative to pipe material, meter}

\item{...}{aditional parameters}
}
\description{
\code{subunit_diameter} calculates the diameter of the manifold line to
achieve a given variation in flow rate
}
\examples{
args_subunit_diameter <- list(
  q_var_subunit = 10,
  h_fim_manifold = 10, s_manifold = 6, s_ini_manifold = 10, n_manifold = 8, dec_manifold = 0.02,
  d_lateral = 0.016, s_lateral = 1, s_ini_lateral = 4, n_lateral = 40, dec_lateral = -0.01,
  coef_em = 6.41e-7, exp_em = 0.54, rc = 1e-4)
do.call(subunit_diameter, args_subunit_diameter)
}
